+++ scripts/sort_weights.py
@@ -50,6 +50,7 @@ def write_supported_weights():
         "BiRefNet": weights_manifest.get_weights_by_type("BIREFNET"),
         "Style models": weights_manifest.get_weights_by_type("STYLE_MODELS"),
         "DepthAnything": weights_manifest.get_weights_by_type("DEPTHANYTHING"),
+        "FBCNN (Jpeg artifact removal)": weights_manifest.get_weights_by_type("FBCNN"),
         "Anyline": helpers.ComfyUI_Anyline.models(),
         "AnimateDiff": weights_manifest.get_weights_by_type("ANIMATEDIFF_MODELS"),
         "AnimateDiff LORAs": weights_manifest.get_weights_by_type(